+++ src/core/core.cpp
@@ -4,7 +4,7 @@
 
 
 #include "../object.h"
-#include "../conv.hpp"
+#include "../traits.hpp"
 #include "conversions.h"
